fix : 저장하지 않은 숏스, 뉴스를 제거하는 행위의 검증 로직 추가

+++ shorts-domain/src/main/kotlin/com/mashup/shorts/domain/membernews/MemberNewsDelete.kt
@@ -2,6 +2,8 @@ package com.mashup.shorts.domain.membernews
 
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import com.mashup.shorts.common.exception.ShortsBaseException
+import com.mashup.shorts.common.exception.ShortsErrorCode
 import com.mashup.shorts.domain.member.Member
 import com.mashup.shorts.domain.news.NewsRepository
 
@@ -14,6 +16,14 @@ class MemberNewsDelete(
 
     fun deleteMemberNews(member: Member, newsIds: List<Long>) {
         val newsList = newsRepository.findAllById(newsIds)
+
+        if (memberNewsRepository.findAllById(newsIds).size == 0) {
+            throw ShortsBaseException.from(
+                shortsErrorCode = ShortsErrorCode.E404_NOT_FOUND,
+                resultErrorMessage = "저장되지 않은 뉴스를 삭제할 수 없습니다."
+            )
+        }
+
         memberNewsRepository.deleteAllByMemberAndNewsIn(member, newsList)
     }
 }
